requirements.txt
Flask>=2.0
Flask-Login>=0.6
Flask-SQLAlchemy>=2.5
Flask-JWT-Extended>=4.4
python-dotenv>=0.21
matplotlib>=3.5
numpy>=1.23 